requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
reportlab==4.0.7
Pillow==11.3.0
aiofiles==23.2.1